change receipt form and receipt field into seperate component, issues on passing fields and watch component
